requirements.txt
fastapi
uvicorn
google-genai
instructor
pydantic
python-dotenv
tenacity
pandas